Normalize IAM Action and principals
Action entry in the policy statement can be a single string or array of
strings in addition to the AWS principal.
