+++ aws/mocks/mock_sns.go
@@ -19,9 +19,7 @@ var defaultSnsPolicy = `
           "*"
         ]
       },
-      "Action": [
-        "SNS:Subscribe"
-      ],
+      "Action": "SNS:Subscribe",
       "Resource": "arn:aws:sns:us-east-1:000000000000:test-topic",
       "Condition": {
         "StringEquals": {
