+++ deployer/template/aws_serverless_function_events.go
@@ -12,6 +12,7 @@ import (
 	"github.com/awslabs/goformation/cloudformation"
 	"github.com/awslabs/goformation/cloudformation/resources"
 	"github.com/coinbase/fenrir/aws"
+	"github.com/coinbase/fenrir/aws/iam"
 	"github.com/coinbase/step/aws/s3"
 	"github.com/coinbase/step/utils/to"
 )
@@ -128,22 +129,6 @@ func ValidateSQSEvent(projectName, configName string, event *resources.AWSServer
 	return hasEventTags(projectName, configName, tags)
 }
 
-type PolicyDocument struct {
-	Version   string
-	Statement []StatementEntry
-}
-
-type StatementEntry struct {
-	Effect    string
-	Action    []string
-	Resource  string
-	Principal PrincipalEntry
-}
-
-type PrincipalEntry struct {
-	AWS interface{}
-}
-
 func ValidateSNSEvent(projectName, configName string, roleArn string, event *resources.AWSServerlessFunction_SNSEvent, snsc aws.SNSAPI) error {
 	// event.Topic is ARN e.g. arn:aws:sns:us-east-1:000000000000:test-topic
 	region, account, resource := to.ArnRegionAccountResource(event.Topic)
@@ -158,7 +143,7 @@ func ValidateSNSEvent(projectName, configName string, roleArn string, event *res
 		return err
 	}
 
-	var policy PolicyDocument
+	var policy iam.PolicyDocument
 	err = json.Unmarshal([]byte(*out.Attributes["Policy"]), &policy)
 	if err != nil {
 		return err
@@ -168,7 +153,8 @@ func ValidateSNSEvent(projectName, configName string, roleArn string, event *res
 	// as the role is specified in the topic's access policy principals.
 	for _, entry := range policy.Statement {
 		valid := false
-		for _, action := range entry.Action {
+
+		for _, action := range entry.NormalizedAction() {
 			if strings.ToLower(action) == "sns:subscribe" {
 				valid = true
 			}
@@ -178,23 +164,7 @@ func ValidateSNSEvent(projectName, configName string, roleArn string, event *res
 			continue
 		}
 
-		var principals []string
-
-		switch entry.Principal.AWS.(type) {
-		case string:
-			principals = []string{entry.Principal.AWS.(string)}
-		case []interface{}:
-			principalInters := entry.Principal.AWS.([]interface{})
-
-			for _, inter := range principalInters {
-				principal, ok := inter.(string)
-				if ok {
-					principals = append(principals, principal)
-				}
-			}
-		}
-
-		for _, principal := range principals {
+		for _, principal := range entry.Principal.NormalizedAWS() {
 			if roleArn == principal {
 				return nil
 			}
